+++ src/Hledger/Flow/Import/Types.hs
@@ -6,16 +6,20 @@ import Prelude hiding (FilePath, putStrLn)
 import Hledger.Flow.Types
 
 data ImportOptions = ImportOptions { baseDir :: FilePath
+                                   , hledgerInfo :: HledgerInfo
                                    , verbose :: Bool
                                    , showOptions :: Bool
                                    , sequential :: Bool }
   deriving (Show)
 
 instance HasVerbosity ImportOptions where
-  verbose (ImportOptions _ v _ _) = v
+  verbose (ImportOptions _ _ v _ _) = v
+
+instance HasSequential ImportOptions where
+  sequential (ImportOptions _ _ _ _ sq) = sq
 
 instance HasBaseDir ImportOptions where
-  baseDir (ImportOptions bd _ _ _) = bd
+  baseDir (ImportOptions bd _ _ _ _) = bd
 
 data ImportDirs = ImportDirs { importDir  :: FilePath
                              , ownerDir   :: FilePath